requirements.txt
absl-py==1.1.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
cvzone==1.5.6
cycler==0.11.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
mediapipe==0.8.8
numpy==1.21.6
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
protobuf==3.20.1
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
requests==2.28.0
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
